requirements.txt
flask==3.1.0
flask-cors==5.0.1
openai==1.65.5
python-dotenv==1.0.1
gunicorn==21.2.0
